Nearly Finished, just need to complete visuals
Program Freezes when "std::cout" is introduced. I'm not sure why this happens, but it's not mandatory for me to be using "std::cout" but I would like to fix this issue regardless.

The fuzzification and defuzzification works, I just need to complete the visuals and fix the issue stated above and I should be good.